Refactor code for improved readability and functionality
- Updated BringStatementGenerationStrategy to handle file path replacements for Linux, macOS, and Windows.
- Rearranged STATIC_LINKING_LIBRARIES in LibUtils for better organization.
- Enhanced CommandManager to streamline command generation and improve logging.
- Modified AOTCompiler utility functions for better file handling and logging.
- Adjusted ObjectFile class to ensure consistent object file naming.
- Cleaned up CompilerUtils by removing unused functions and improving logging.
- Added additional diagnostic warnings in diagnostic_push.h for better code safety.
- Simplified Interpreter logic by removing unnecessary return statements.
- Improved logging in Utils for better debugging and file handling.
- Updated test cases to use lambda for character transformation for better compatibility.

+++ src/IR/irGen/Types/FlowWingClass.h
@@ -249,6 +249,6 @@ public:
   }
 
   inline auto addMemberVariableTypeName(std::string name) -> void {
-    _memberVariablesTypeNameMap.insert({name, 1});
+    _memberVariablesTypeNameMap.insert({name, static_cast<int8_t>(1)});
   }
 };
